% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatting.R
\name{.isKnownField}
\alias{.isKnownField}
\title{Detect whether a given field number is defined for a specified global
message type.  Garmin (and maybe others) include many fields that are not
documented in the FIT Profile}
\usage{
.isKnownField(field_definition_number, global_message_number)
}
\arguments{
\item{field_definition_number}{integer of length 1. This is the field
number we're checking the existence of in the FIT specification.}

\item{global_message_number}{integer of length 1. Specifies the global
message number for the message type we're looking at.}
}
\description{
Detect whether a given field number is defined for a specified global
message type.  Garmin (and maybe others) include many fields that are not
documented in the FIT Profile
}
\keyword{internal}
