% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitFile_accessors.R
\name{FitFile-accessors}
\alias{FitFile-accessors}
\alias{getMessageTypes}
\alias{getMessageTypes,FitFile-method}
\alias{getMessagesByType}
\alias{getMessagesByType,FitFile,integer-method}
\alias{getMessagesByType,FitFile,character-method}
\alias{records}
\alias{records,FitFile-method}
\alias{laps}
\alias{laps,FitFile-method}
\title{Extracting messages from Fit Files}
\usage{
getMessageTypes(fitFile)

\S4method{getMessageTypes}{FitFile}(fitFile)

getMessagesByType(fitFile, message_type)

\S4method{getMessagesByType}{FitFile,integer}(fitFile, message_type)

\S4method{getMessagesByType}{FitFile,character}(fitFile, message_type)

records(fitFile)

\S4method{records}{FitFile}(fitFile)

laps(fitFile)

\S4method{laps}{FitFile}(fitFile)
}
\arguments{
\item{fitFile}{A \code{\link{FitFile-class}} object.}

\item{message_type}{Either an integer or character vector (length 1),
specifying either a global message number or message type respectively.}

\item{global_message_number}{Integer specifying the message number of
the message type wanted.}
}
\value{
Either a \code{tibble} or a \code{list} of \code{tibble}. See \code{details} for more
information.
}
\description{
Methods for retrieving messages from \code{\link{FitFile-class}} objects.
}
\details{
The FIT file specification allows for a large number of message types.
FitFileR provides accessor methods for some of the most common. These
include \code{records()} and \code{laps()}.

If a predefined function doesn't exist for the message type you want to
extract, any message type can be retrieved with \code{getMessagesByType()}.
The second argument can take either the global message number (as specified
in the FIT File definition) of message type you want, or the message name.
A list of names for the message types held in a \code{\link{FitFile-class}}
object can be retrieved with \code{\link{getMessageTypes}}.

The return type is dependant upon whether the \code{\link{FitFile-class}}
contains multiple message definitions for the same message type.  It is not
uncommon for this to occur e.g. if a new sensor is added during an activity
the \emph{records} field definition will change.  If there is a single definition
for the message type a \code{tibble} will be returned, otherwise a \code{list} of
\code{tibble}s is returned.  The length of this \code{list} reflects the number of
unique definitions for the message type within the file.  It may be
straightforward to combine these \code{tibbles} e.g. via \code{\link{rbind}}, but
this is left to the user.
}
