% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/methods_FitFile_dim.R
\docType{class}
\name{FitFile-class}
\alias{FitFile-class}
\alias{length,FitFile-method}
\title{An S4 class representing a FIT file}
\usage{
\S4method{length}{FitFile}(x)
}
\arguments{
\item{x}{An object of class \code{FitFile}.}
}
\description{
An S4 class representing a FIT file
}
\section{Slots}{

\describe{
\item{\code{header}}{A list containing details of the file header}

\item{\code{messages}}{A list of \linkS4class{FitDataMessage}s}
}}

