% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITfileR.R
\docType{package}
\name{FITfileR}
\alias{FITfileR}
\title{FITfileR: A package reading FIT files using native R code.}
\description{
Provides functionality for reading FIT (Flexible and Interoperable Data
Transfer) files without any dependence on external software or SDKs.
}
