% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FitMessageHeader-class}
\alias{FitMessageHeader-class}
\title{The \code{FitMessageHeader} class represents the single-byte header that
precedes all FIT messages.  It forms a part of all
\code{FitDefinitionMessage} and \code{FitDataMessage} instances.}
\description{
The \code{FitMessageHeader} class represents the single-byte header that
precedes all FIT messages.  It forms a part of all
\code{FitDefinitionMessage} and \code{FitDataMessage} instances.
}
\section{Slots}{

\describe{
\item{\code{is_defintion}}{Logical indicating if this is a definition message.}

\item{\code{has_developer_data}}{Logical defining whether the message contains
developer data.}

\item{\code{local_message_number}}{The 'local' message number for this message
type.}

\item{\code{time_offset}}{Numeric (length 1) giving the time offset for messages
with compressed time stamps.}

\item{\code{raw_rep}}{Raw (length 1) matching the single byte that defined this
message header.  Used internally as a quick comparison to see if this is the
same header as the previous message to speed up file reading.}
}}

