% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FitMessages}
\alias{FitMessages}
\alias{FitDefinitionMessage-class}
\alias{FitDataMessage-class}
\alias{FitDataMessageWithDevData-class}
\title{S4 classes representing various aspects of FIT messages.}
\description{
S4 classes representing various aspects of FIT messages.

The \code{FitDefinitionMessage} class represents definition messages.

The \code{FitDataMessage} class holds data messages.

The \code{FitDataMessageWithDevData} class extends \code{FitDataMessage}
with additional slots to store the definitions of developer data fields that
are not defined in the standard FIT file specification.
}
\section{Slots}{

\describe{
\item{\code{global_message_number}}{A integer of length 1.}

\item{\code{field_definition}}{A data.frame.}

\item{\code{messages}}{A data.frame.}

\item{\code{header}}{An object of class \code{FitMessageHeader}.}

\item{\code{definition}}{An object of class \code{FitDefinitionMessage}.}

\item{\code{fields}}{A list containing the data encoded in this message.}
}}

