% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fit.R
\name{ReadFit}
\alias{ReadFit}
\title{Read a FIT file}
\usage{
ReadFit(fileName, verbose = FALSE)
}
\value{
`ReadFit()` returns an object of class `FitData`.
}
\description{
Read a FIT file
}
\author{
[Martin R. Smith](https://orcid.org/0000-0001-5660-1727) 
(<martin.smith@durham.ac.uk>)
}
