% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readFitFile}
\alias{readFitFile}
\title{Read a FIT file}
\usage{
readFitFile(fileName)
}
\arguments{
\item{fileName}{A character specifying the FIT file to be read.}
}
\value{
An object of class \code{[FitFile-class]}
}
\description{
Reads a specified FIT file and returns an object of class \code{FitFile}
}
\examples{
garmin_file <- system.file("extdata", "Activities", "garmin-edge530-ride.fit",
                           package = "FITfileR")
garmin <- readFitFile(garmin_file)

}
