% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_type_checks.R
\name{readFitFileType}
\alias{readFitFileType}
\title{Extract the type of FIT file}
\usage{
readFitFileType(fileName)
}
\arguments{
\item{fileName}{A character specifying the FIT file to be read.}
}
\value{
As character vector of length 1.  This will contain the text
representation of the file type e.g. 'activity'
}
\description{
Quickly determine what type of FIT file you have without decoding all the
data.  Typically types include 'activity', 'workout' and 'course'.  The full
range of options is defined in the FIT file specification.
}
\examples{
garmin_file <- system.file("extdata", "Activities", "garmin-edge530-ride.fit",
                           package = "FITfileR")
garmin <- readFitFileType(garmin_file)

}
