% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitFile_accessors.R
\name{FitFile-accessors}
\alias{FitFile-accessors}
\alias{listMessageTypes}
\alias{listMessageTypes,FitFile-method}
\alias{getMessagesByType}
\alias{getMessagesByType,FitFile,integer-method}
\alias{getMessagesByType,FitFile,character-method}
\alias{messages,FitFile-method}
\alias{file_id}
\alias{file_id,FitFile-method}
\alias{records}
\alias{records,FitFile-method}
\alias{laps}
\alias{laps,FitFile-method}
\alias{events}
\alias{events,FitFile-method}
\alias{hrv}
\alias{hrv,FitFile-method}
\title{Extracting messages from FIT Files}
\usage{
listMessageTypes(fitFile)

\S4method{listMessageTypes}{FitFile}(fitFile)

getMessagesByType(fitFile, message_type)

\S4method{getMessagesByType}{FitFile,integer}(fitFile, message_type)

\S4method{getMessagesByType}{FitFile,character}(fitFile, message_type)

\S4method{messages}{FitFile}(fitFile)

file_id(fitFile)

\S4method{file_id}{FitFile}(fitFile)

records(fitFile)

\S4method{records}{FitFile}(fitFile)

laps(fitFile)

\S4method{laps}{FitFile}(fitFile)

events(fitFile)

\S4method{events}{FitFile}(fitFile)

hrv(fitFile)

\S4method{hrv}{FitFile}(fitFile)
}
\arguments{
\item{fitFile}{A \code{\link{FitFile-class}} object.}

\item{message_type}{Either an integer or character vector (length 1),
specifying either a global message number or message type respectively.}
}
\value{
Either a \code{tibble} or a \code{list} of \code{tibble}. See \code{details} for more
information.
}
\description{
Methods for retrieving messages from \code{\link{FitFile-class}} objects.
}
\details{
The FIT file specification allows for a large number of message types.
FITfileR provides accessor methods for some of the most common. These
include \code{records()} and \code{laps()}.

If a predefined function doesn't exist for the message type you want to
extract, any message type can be retrieved with \code{getMessagesByType}.
The second argument can take either the global message number (as specified
in the FIT File definition) of message type you want, or the message name.
A list of names for the message types held in a \code{\link{FitFile-class}}
object can be retrieved with \code{listMessageTypes}.

The return type is dependant upon whether the \code{\link{FitFile-class}}
contains multiple message definitions for the same message type.  It is not
uncommon for this to occur e.g. if a new sensor is added during an activity
the \emph{records} field definition will change.  If there is a single definition
for the message type a \code{tibble} will be returned, otherwise a \code{list} of
\code{tibble}s is returned.  The length of this \code{list} reflects the number of
unique definitions for the message type within the file.  It may be
straightforward to combine these \code{tibbles} e.g. via \code{\link{rbind}}, but
this is left to the user.
}
