% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{FitFile-class}
\alias{FitFile-class}
\title{An S4 class representing a FIT file}
\description{
An S4 class representing a FIT file
}
\section{Slots}{

\describe{
\item{\code{header}}{A list containing details of the file header}

\item{\code{messages}}{A list of \linkS4class{FitDataMessage}s}
}}

