% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFileR.R
\docType{package}
\name{fitFileR}
\alias{fitFileR}
\alias{fitFileR-package}
\title{fitFileR: A package reading .fit files using native R code.}
\description{
fitFileR: A package reading .fit files using native R code.
}
