% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFileR.R
\docType{package}
\name{fitFileR}
\alias{fitFileR}
\title{fitFileR: A package reading FIT files using native R code.}
\description{
Provides functionality for reading FIT (Flexible and Interoperable Data 
Transfer) files without any dependence on external software or SDKs.
}
